pyproject.toml
[tool.poetry]
name = "whatsapp_bot"
version = "1.0.0"
description = ""
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
firebase-admin = "^4.4.3"
pydantic = "^1.9.0"
aiohttp = "^3.8.1"
google-generativeai = "^0.3.1"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.4"
pylint = "^2.10.2"
black = "^21.12b0"
isort = "^5.10.1"

[tool.poetry.scripts]
dev = "run:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

